fixed bugs

+++ controller/cart.controller.js
@@ -1,9 +1,10 @@
 const Cart = require('../models/Cart')
+const {sendResponseError} = require('../middleware/middleware')
 
 const getCartProducts = async (req, res) => {
   try {
     const carts = await Cart.find({userId: req.user._id}).populate('productId')
-    // console.log(carts)
+  console.log("carts",carts)
     res.status(200).send({status: 'ok', carts})
   } catch (err) {
     console.log(err)
@@ -28,11 +29,14 @@ const addProductInCart = async (req, res) => {
 }
 const deleteProductInCart = async (req, res) => {
   try {
-    await Cart.findByIdAndRemove(req.params.id)
+    if(req.params.id !== undefined){
+      console.log("req.params.id",req.params.id);
+      await Cart.findByIdAndRemove(req.params.id)
+    }
     res.status(200).send({status: 'ok'})
   } catch (e) {
     console.log(e)
-    sendResponseError(500, `Error ${err}`, res)
+    sendResponseError(500, `Error ${e}`, res)
   }
 }
 module.exports = {addProductInCart, deleteProductInCart, getCartProducts}
